Krushkal's Algorithm
A basic Implementation Of Krushkal's Algorithm to find the minimum spanning tree of a graph in Cpp using structure. we have taken inputs like no. of nodes., the weight of each edge from its starting node to its terminating node from user and then print the minimum cost of constructing the tree.